+++ src/types/core/permission.ts
@@ -1,0 +1,10 @@
+import { ApplicationCommandPermissionData } from "discord.js";
+
+export type IPermissions =
+  | ApplicationCommandPermissionData
+  | ApplicationCommandPermissionData[]
+  | (() =>
+      | ApplicationCommandPermissionData
+      | ApplicationCommandPermissionData[]
+      | Promise<ApplicationCommandPermissionData>
+      | Promise<ApplicationCommandPermissionData[]>);